feat(release-please.yml): add release-please workflow to automate versioning and releases
This commit adds a new workflow file that uses the release-please-action to automate versioning and releases. The workflow is triggered on push events to the main branch and runs on ubuntu-latest. The release-please-action is configured to bump patch for minor pre-major and release a new version of the package named release-please-action.